Supervisor Loss for Deterministic Policies (#3472)
Initial draft of case 1, 2, and 4 in Matt's document. Features
supervised loss against a deterministic teacher's output. The loss uses
binary cross entropy as well as MSE and cosine similarity for testing.
There are a bunch of config parameters to make it easy to test various
paths.

To use, change the loss config in your recipe to be something like
{"action_supervised": ActionSupervisedConfig()}.

It's a a first pass:
- Still need to integrate extracting teacher action tokens from the env.
- Testing should collapse the loss options down to one.
- There are a bunch of hypers that need to be tuned.
- Matt's document wants much more than this. Those changes will modify
this file's logic, particularly where we call the student policy's
forward. For now, the current implementation is a faster way to run a
supervised loss.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211789814654235)

---------

Co-authored-by: Richard Higgins <richard@relh.net>