Fix cron job to update PR embedding cache (#3493)
- [x] Need to have cron job checkout main before marking PR as ready!

Two sets of changes
- Various improvements to the script that actually updates the PR
embedding cache
- Fix some things in the cron job helm file to make sure the correct
script is run and the full git history is actually available to the
script. And importantly make sure to use `uv run` with `--no-sync` to
make sure that we don't install hundreds of large dependencies that we
never need.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211801244346349)