Action supervised loss updates (#3558)
It's now complete, reads teacher actions, and runs on mettabox without
error. I have not tested whether it trains a good policy or not because
I now need to activate one of the good teacher policies.

I left extensive notes on how to use the branches and where I think
we'll want to tweak things if training doesn't work. Finally, we'll want
to edit this further once we want to work in PPO in addition to the
supervised loss (as discussed: we want to run supervised first,
checkpoint, then run PPO so this delivers on that).

Finally, there was some confusion: initially the teacher was to deliver
a binary, multi-class target. However, it now delivers one-hot target.
As such, I switched from BCE to CE loss.

To use, set the following in your train tool
`losses=LossConfig(loss_configs={"action_supervised":
ActionSupervisedConfig()})`

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211843294506212)