Update chests to work on vibes (#3537)
This switches chests from operating on position to operating on vibes.
The current default is that you put all your stuff in the chest as the
default action, and then vibe a specific resource to get that resource
out. I.e., you vibe germanium to get germanium -- just like you do at
the extractor! We're so clever.

Happy to play with that the actual vibing should be, but I'm pretty sure
this is the right mechanic for now.


https://app.asana.com/1/1209016784099267/project/1209096222434381/task/1211829872747460?focus=true

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211833621304891)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>