clear metta/rl/loss/__init__.py (#3597)
Removes an **init** that handled lazy loading, but resulted in
complexity. Post this removal, imports are a little uglier, but more
explicit.
  
The underlying purpose of this is to reduce complexity, since I / we
have been hit recently by circular dependencies, and having complexity
in that space makes it harder to clean up circular dependencies when
they happen.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211868931249046)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>