Run `install.sh` in skypilot job setup (#3652)
Jack ran into a problem where a default profile was not discovered in
skypilot run setup. This adds that

Separately, the docker image that runs each job will already have run
`./devops/tools/install-system.sh`. But in the case that it did so on
stale dependencies, and we have not yet published a new docker image, we
want the skpyilot runner to also run it. `./install.sh` is the preferred
entrypoint, and so we use it here. The docker image doesn't do so
because it doesn't install uv deps in frozen mode, but we are about to
run a command that would do the same in the skypilot job setup anyway,
so it is safe to do here.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211906675500345)