Various cooling-themed updates to remote eval tools (#3660)
- Renamed and added inline comments to the various eval Tool
implementations. Should now be clear which is used for executing an eval
locally, executing an eval within the context of a remote worker, and
requesting a remote eval
- Enforced that remote evals specify a eval_Task_id and have and can
authenticate with a stats_server_uri -- if they dont, then theres no
point in running them
- Moved RemoteJobTool to ToolWithResult, since that's closer to what it
really is, and put it in the same file as the Tool definition. We could
consider instead making this part of run.py at a later point
- push_metrics_to_wandb is default True now
- Defined the Tool that remote eval workers use inline with the recipe
they specify a path to; they are the only caller
- Simplified error handling/catching for ToolWithResult
- Moved the recipe that gets specified for remote eval workers into
`experiments.remote_evals` and out of `experiments.evals`, since the
latter contains specific evaluation configs -- a different type of thing

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211914921562304)