fix chest observations (#3693)
We previously had feature_ids for inventory items done in Inventory, and
feature_ids for input / output done in the ObservervationEncoder. But we
were failing to initialize Inventories with a list of feature_ids, so
their get_feature_id field consistently returned zero.

This moves all of the getting-of-feature-ids to ObservationEncoder, and
removes a bunch of fallback behavior that was hiding the issue.


https://app.asana.com/1/1209016784099267/project/1209096222434381/task/1211917181115612?focus=true

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211928552917357)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>