Remove swap / swappable (#3698)
This removes the swap action and surrounding infrastructure.

Swap has been effectively disabled for a while. The version we had only
made sense in the context of orientation, which we'd previously removed.

If we want to have swapping in the future, it should be crammed into the
`onUse` of the swappable objects, and should involve vibing or
something.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211928795671999)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>