Remove assembler exhaustion mechanic (#3720)
I added this mechanic as a way for assemblers (/ extractors) to become
worse over time as they were used. We haven't ended up using this --
we've just used max_uses. I don't think we need two forms of "wears
out", and this currently depends on floating point math. So removing,
with an understanding that we can add it back pretty easily if we decide
we want something like this in the future.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211940906723340)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>