Update MISSION, tweaks some parameters (#3855)
Updated the MISSION.md documentation and station parameters to reflect
new game mechanics, including vibe-based interactions and revised
resource values.

This also tweaks some of the parameters we have that seemed off when I
was updating the docs:
* reverts some germanium changes we had that made germanium extraction /
usage not scale cleanly with the number of agents.
* reduces energy usage for extracting silicon, mostly because it felt
overly punishing, and because the amount of generated silicon per load
was recently reduced; so this slightly corrects the amount of energy
needed to get a "full load".

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211986241377303)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>