PPO and ActoionSupervised losses refactor (#3834)
# Overview
This splits the PPO loss implementation into two separate losses, one
for the actor and one for the critic, with the motivation of allowing
other losses (eg ActionSupervised) to solely use the PPO value loss.
Actor retains the entropy loss term (literally one line but can be
turned off by simply setting the coef to 0). I updated Losses to default
to this implementation but left the original PPO implementation in there
for use cases where we want everything in one file. We risk the two
implementations drifting apart from each other.

The major challenge was splitting the two terms cleanly while still
leaving prio sampling and v-trace intact. For instance, v-trace requires
importance sampling ratios which is computed at the actor but also
requires recomputing advantage which, in turn, requires knowing GAE
gamma and lambda coefficients. However, those coefficients are set
elsewhere, in ppo_critic.py. To avoid needing to set GAE gamma and
lambda coefficients in _both_ locations, I pass the coefficients from
ppo_critic to ppo_actor via the shared_loss_data buffer. It's admittedly
a bit awkward.

Also fixed a bug found during the refactor. When prio_alpha is set to 0
as it currently is to turn off prioritized replay sampling, we still
sample with replacement. This means that each minibatch after the 0th
has a chance of resampling a previously-trained-on trajectory. Over the
32 minibatches we use on mettabox and given the minibatch size, this
means we ignore an average of 37% of our trajectories.

These have been tested on mettabox: first implemented the big fix above
on our original, single-file ppo implementation to generate a training
curve to 6b steps on abes and determined that it met or exceeded the
original ppo implementation. Then tested the split-file implementation
against the single-file, bug-fixed implementation and the training
profile is equivalent. SPS is roughly the same.

Finally, also refactored the ActionSupervised loss to be a lot simpler
and more set up for long-term functionality (it was originally more
prototype-oriented) as its now ready to integrate fully with a
concurrent PPO loss. I only debugged this by ensuring it trains locally.

# Usage
You don't have to do anything, the changes are now the default.

To revert to the original, single-file PPO loss, do this in your
recipe's train tool:
```
    losses_cfg = LossesConfig()
    losses_cfg.ppo_actor.enabled = False
    losses_cfg.ppo_critic.enabled = False
    losses_cfg.ppo.enabled = True
    trainer_cfg = TrainerConfig(losses=losses_cfg)
```

To run the action supervised loss while also using PPO's value loss to
train the policy's critic, do this:
```
    losses_cfg = LossesConfig()
    losses_cfg.ppo_actor.enabled = False
    losses_cfg.supervisor.enabled = True
    losses_cfg.ppo_critic.rollout_forward_enabled = False # we want the ActionSupervised loss to handle the rollout
    losses_cfg.ppo_critic.sample_enabled = False # want ActionSupervised to handle replay sampling during training
    losses_cfg.ppo_critic.train_forward_enabled = False # want ActionSupervised to handle forwarding policy in training
    trainer_cfg = TrainerConfig(losses=losses_cfg)
```

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211989650101303)