Remove a semi-stale doc (#3905)
This doc describes the impact of updating actions. Some aspects are
still relevant, but (e.g.) references to a multi-discrete action space
no longer are. I'd be happy if you want to update the doc, but I'm
guessing we can purge it and recreate later as needed.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211998172989785)

---------

Co-authored-by: S. Alex Smith <sasmith@stem.ai>