Enable submitting metta-trained policies to cogames leaderboard (#4024)
This PR adds support for specifying the init-kwargs part of PolicySpec
in `cogames submit`. Doing so lets us pass a checkpoint uri to
CheckpointPolicy, which allows us to hydrate a metta-trained policy via
`initialize_or_load_policy`, which allows such policies to be submitted
to the leaderboard

Adds an example to the `recipes/.../cogs_v_clips.py` file for how to do
so:

```
uv run cogames submit \
  -p metta.rl.checkpoint_manager.CheckpointPolicy \
  -k checkpoint_uri=s3://softmax-public/policies/...:v1.mpt \
  -n your-policy-name-for-leaderboard \
  --skip-validation
```

For now we need to run --skip-validation because cogames validation
doesn't assume the leaderboard runners get to run with the `metta` repo
available, but in practice they do. We should pickle more from the input
class path instead of this in the future

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1212085964638476)