Remove object_type_name usage from miniscope; remove stale files (#4023)
This is a push towards removing type_ids, and clarifying what type_name
means. E.g., is it the name of a config, or how the object is rendered,
or something else?

In this PR, we simplify the problem by removing superfluous uses of
object_type_name. Some of this is in miniscope, some is in map_previews
(which appear to be fully unused), and some is in
`01-hello-world-marimo.py`, which I've had to update a lot and I think
is not used, and so I'm deleting it. If it is used, people should
complain!

Hopefully this + similar work will let us stop requiring the sim to
expose object_type_names. It would be better if they were read from the
config, if they were needed at all.

Co-authored-by: S. Alex Smith <sasmith@stem.ai>