Remove some floats (#4058)
Removed the resource loss probability feature from MettaGrid and
optimized episode completion percentage calculation.

This is part of removing floats from the simulation. For features that
are in active use, I'm aiming to replace floating calculations with
integer calculations. For features that aren't in active use, I'm just
removing the feature.


https://app.asana.com/1/1209016784099267/project/1209096222434381/task/1211547145607801?focus=true

Co-authored-by: S. Alex Smith <sasmith@stem.ai>