Update nim/nimby installation in bootstrap.py: (#4137)
Fixes:
- Only add symlinks to an existing dir in PATH if nim/nimby are not
already findable. this stops us from directly overwriting things already
in ~/.nimby with symlinks to itself
- Distinguish between (a) target directories to install into (if in
path), and (b) directories to add to look for binaries in. And remove
nimby from the former. This was causing uv to get installed in that dir,
and nimby symlinks to overwrite existing entries in that dir
- More accurate install messages: clarify which of nim/nimby are being
installed/linked/etc, and if it's because they're out of date or not. We
previously had some misleading messages
- nimby use` overwrites the ~/.nimby... bin, so we now try to install
nimby to a different target loc, and then dont need to symlink later.
This prevents `nimby use` installing nim but breaking nimby

Cooling:
- Remove fallback version-checking logic, refactor version-checking code
to be more DRY
- Refactor out installing nimby into a tmp dir; separable step from
installing nim
- Remove a bunch of redundant verification that only preempts a more
explicit failure at a later point

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>