Remove bad variant (#4160)
CogToolsOnlyVariant makes all tools created with only the `gear` vibe,
but this results in multiple protocols with exactly the same numbers of
agents / vibes, and this makes it impossible the agents to indicate
which protocol they want to use.

Also, intuitively, this doesn't seem like a particularly valuable
variant. So I'm scrapping it.

Let me know if I'm missing something.

Co-authored-by: S. Alex Smith <sasmith@stem.ai>