Add a callback for inventory updates (#4197)
Implemented an inventory change notification system using a callback
mechanism to make it harder to miss stat updates.

The point of this PR is 50% "homogenize how things work" and 50% "make
it so I can have functions that just take and operate on inventories".
And this is all in the realm of "make it so Assemblers can work with
Agents and / or Chests".

- Added an `on_inventory_change` virtual method to `HasInventory` class
that gets called whenever an inventory changes
- Modified `Inventory` class to store a reference to its owner and call
the callback when inventory changes
- We could generalize this to a list and have this be a more standard
"register your callback" sort of thing, but no clear need to.


https://app.asana.com/1/1209016784099267/project/1209096222434381/task/1211405131768007?focus=true

Co-authored-by: S. Alex Smith <sasmith@stem.ai>