By default, evaluate only remotely if in skypilot run, and only locally if not (#4215)
- Trainer config detects if we're on skypilot, by default, evaluates
only remotely if so, and only locally if not
- Various recipes had custom overriding behavior for evaluate_local and
evaluate_remote. Removed some of that
- Stable and CI now explicitly either do not run eval in their training
jobs or only do so locally; not yet testing remote evals. This maintains
previous behavior with this switch in defaults
- Fixed all evaluate recipes to require policy_uris; some previously had
it as an optional arg, but those were destined to fail
- Fixed an instance of passing behavior cloning policy uri incorrectly
in a training recipe; probably the result of bad merge resolution at
some point
- recipes.experiment.remote_eval uses default stats server uri instead
of hardcoding it to prod. Only matters for local testing of remote eval
runners