Add a variant where Assemblers pull from nearby chests (#4224)
Added a new mission variant where assemblers can draw resources directly
from specialized resource chests.

In this case, cogs can't withdraw from the chests. This is partially
because that's easier, and partially because they don't need to. I think
a more full version could have a "insert" and "withdraw" emoji or
something, but this version is mostly for people trying it out.

So far, seems good.

Co-authored-by: S. Alex Smith <sasmith@stem.ai>